Added media queries

+++ controllers/usersController.js
@@ -60,7 +60,8 @@ module.exports.update = async function (req, res) {
                     {
                         if(user.avatar)
                         {
-                            fs.unlinkSync(path.join(__dirname,'..',user.avatar));
+                            if(fs.existsSync(path.join(__dirname,'..',user.avatar)))
+                                fs.unlinkSync(path.join(__dirname,'..',user.avatar));
                         }
                         user.avatar = User.avatarPath+'/'+req.file.filename;
